Rewrite PKSM Bridge protocol to support dynamic file size.
This commit implements the PKSM Bridge Protocol (v1), which is a simple
protocol built on top of the TCP connection to allow for future changes
through versioning and transmitting dynamic file sizes.

NOTE: The implementation shares a lot of code with the PKSM
implementation. We should refactor these into a common dependency.